requirements.txt
numpy
scipy
scikit-learn
scikit-image
opencv-python
Pillow
fonttools
cairosvg
torch
torchvision
tqdm
pyyaml
